Support multiple CTC ports + clean up
